각 entity들의 pk타입을 Integer로 변경

+++ src/main/java/com/JobAyong/entity/InterviewArchive.java
@@ -14,7 +14,7 @@ public class InterviewArchive {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "interview_archive_id")
-    private Long interviewArchiveId;
+    private Integer interviewArchiveId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "email", nullable = false)
